backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Finance Tracker - full-stack personal finance app"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "whitenoise>=6.12.0",
    "dj-database-url>=3.1.2",
    "psycopg2-binary>=2.9.11",
    "gunicorn>=25.3.0",
    "pillow>=12.1.1",
]
